requirements.txt
Django==4.2.16
django-allauth==65.2.0
django-jazzmin==3.0.1
mercadopago==2.2.3
Pillow==11.0.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
stripe==11.6.0
gunicorn==23.0.0
